+++ src/main/java/com/pml/dto/StorageDeviceDTO.java
@@ -22,7 +22,7 @@ public class StorageDeviceDTO implements Serializable {
 	private String description;
 	private boolean itWorks = true; 
 	private boolean itComposed;
-	private Double sizeInMB;
+	private Integer sizeInMB;
 	private Integer architecture;
 	private Integer type;
 	
@@ -47,7 +47,7 @@ public class StorageDeviceDTO implements Serializable {
 	
 	public StorageDeviceDTO(Long id, String patrimonyId, Date createdDate, Date lastModifiedDate,  EquipmentType equipmentType,
 			String manufacturer, String model, String description, Integer sector, boolean itWorks, boolean itComposed, 
-			Double sizeInMB, StorageDeviceArchitecture architecture, StorageDeviceType type) {
+			Integer sizeInMB, StorageDeviceArchitecture architecture, StorageDeviceType type) {
 		this.id = id;
 		this.createdDate = createdDate;
 		this.lastModifiedDate = lastModifiedDate;
@@ -143,11 +143,11 @@ public class StorageDeviceDTO implements Serializable {
 		this.itComposed = isItComposed;
 	}
 
-	public Double getSizeInMB() {
+	public Integer getSizeInMB() {
 		return sizeInMB;
 	}
 
-	public void setSizeInMB(Double sizeInMB) {
+	public void setSizeInMB(Integer sizeInMB) {
 		this.sizeInMB = sizeInMB;
 	}
