+++ src/main/java/com/pml/dto/StorageDeviceNewDTO.java
@@ -16,7 +16,7 @@ public class StorageDeviceNewDTO implements Serializable {
 	private String description;
 	private boolean itWorks = true;
 	@NotNull(message = "This field is mandatory")
-	private Double sizeInMB;
+	private Integer sizeInMB;
 	@NotNull(message = "This field is mandatory")
 	private Integer architecture;
 	@NotNull(message = "This field is mandatory")
@@ -58,11 +58,11 @@ public class StorageDeviceNewDTO implements Serializable {
 		this.itWorks = itWorks;
 	}
 	
-	public Double getSizeInMB() {
+	public Integer getSizeInMB() {
 		return sizeInMB;
 	}
 
-	public void setSizeInMB(Double sizeInMB) {
+	public void setSizeInMB(Integer sizeInMB) {
 		this.sizeInMB = sizeInMB;
 	}
